Move AES functionality to a new file
